// src/models/Movie.ts
export interface Movie {
    Poster: string
    Title: string
    Type: string
    Year: string
    imdbID: string
}

export interface MovieDetails {
    Actors: string
    Awards: string
    Country: string
    Director: string
    Genre: string
    Language: string
    Metascore: string
    Plot: string
    Poster: string
    Rated: string
    Ratings: any
    Released: string
    Response: string
    Runtime: string
    Title: string
    Type: string
    Writer: string
    Year: string
    imdbID: string
    imdbRating: string
    imdbVotes: string
    totalSeasons: string
}